Refactor prefix_sum function to use multithreading
The prefix_sum function in p1.cpp has been refactored to use multithreading for improved performance. The function now utilizes the C++ <thread> API to perform its work using a number of threads matching the thread_count parameter. This refactor does not change the signature of the prefix_sum function.